FIX: Regression on category matrix selections (#90)
When no category ids are present on a global filter,`this.filterTag.category_ids` returns `""`.  When trying to map and split this empty string to convert the possible string values to integers we do

```
.map((str) => Number(str));
```

The problem was that `Number("")` evaluates to `0`. This broke our topic filtering logic as you could not leave the category_ids section blank for a Global Filter, it was always assigning `0` at the minimum. This did not allow for the _Leave every checkbox blank to select all categories_ functionality.